Texts work

+++ db/migrate/20151215093631_create_reports.rb
@@ -1,7 +1,7 @@
 class CreateReports < ActiveRecord::Migration
   def change
     create_table :reports do |t|
-      t.decimal :snow_allday, :snow_day, :snow_night, :high, :low
+      t.float :snow_allday, :snow_day, :snow_night, :high, :low
       t.datetime :date_epoch
       t.belongs_to :resort
